requirements.txt
google-genai>=1.47.0
pydantic>=2.6.0
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
python-dotenv>=1.0.1
manim>=0.18.0
